DIGEQ-7 refactoring test files
